Only load file-based forms when running the devserver
